# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IUT Cafeteria API
  description: |
    API for the IUT Cafeteria ordering system.

    ## Authentication
    All endpoints (except /login, /register, /health) require JWT authentication.
    Pass the token via:
    - `Authorization: Bearer <token>` header
    - `token` httpOnly cookie

    ## Correlation IDs
    All requests are assigned an `x-correlation-id` header for distributed tracing.
  version: 1.0.0
  contact:
    name: IUT Cafeteria Team
    email: admin@iut-cafeteria.app
servers:
  - url: http://localhost:3000
    description: Order Gateway (Development)
  - url: http://localhost:3001
    description: Identity Provider (Development)
  - url: http://localhost:3002
    description: Stock Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    MenuItem:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Chicken Biryani'
        description:
          type: string
          example: 'Fragrant rice with tender chicken'
        price:
          type: number
          format: float
          example: 120.00
        category:
          type: string
          enum: [rice, curry, snacks, drinks, dessert]
        is_available:
          type: boolean
        stock:
          type: integer
          example: 50
...
paths:
...
  /menu:
    get:
      tags: [Menu]
      summary: Get available menu items
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: List of menu items
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MenuItem'
...
  /cafeteria/tokens:
    get:
      tags: [Cafeteria]
      summary: Get booked tokens
      security:
        - BearerAuth: []
        - CookieAuth: []
      responses:
        '200':
          description: Token list
  /cafeteria/tokens/bulk:
    post:
      tags: [Cafeteria]
      summary: Book multiple meal tokens
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tokens]
              properties:
                tokens:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [dinner, iftar]
                      date:
                        type: string
                        format: date
      responses:
        '202':
          description: Tokens booked

  # ─── Reviews ──────────────────────────────────────────────────
...
  /admin/menu:
    get:
      tags: [Admin]
      summary: List all menu items (including unavailable)
      responses:
        '200':
          description: Full menu list

    post:
      tags: [Admin]
      summary: Create menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price, category]
              properties:
                name:
                  type: string
                description:
                  type: string
                price:
                  type: number
                category:
                  type: string
                is_available:
                  type: boolean
                  default: true
                image_url:
                  type: string
                initial_stock:
                  type: integer
                  default: 50
      responses:
        '201':
          description: Item created
  /admin/menu/{id}:
    put:
      tags: [Admin]
      summary: Update menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item updated

    delete:
      tags: [Admin]
      summary: Delete menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Item deleted
  /admin/stock/{itemId}:
    put:
      tags: [Admin]
      summary: Update stock quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity:
                  type: integer
                  minimum: 0
      responses:
        '200':
          description: Stock updated
...
  /stock/{itemId}:
    get:
      tags: [Admin]
      summary: Get stock level
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Stock info
          content:
            application/json:
              schema:
                type: object
                properties:
                  itemId:
                    type: integer
                  quantity:
                    type: integer
                  source:
                    type: string
                    enum: [cache, database]
...